requirements.txt
discord.py
python-dotenv
aiohttp # Async HTTP requests
aiosqlite # Asynchronous SQLite database access
typing # Type hints